Refactor ZoneEditingPage
-uses ZoneStage imported from ZoneStage for separation of concerns
